working better
hacked in fixes to upstream libs couch_cacher and calvad_reducer

The new storage approach in CouchDB is *not* to include the geo data in
the imputation result.  That is, I am not storing the length and the
freeway, for various reasons, but mostly because you can't know what you
need to know when you run the imputation.

So I need to add a bit to the couch_cacher routine to write the
length (and the freeway while I'm at it) to each record pulled from
couchdb (representing the hourly impute result)